+++ spec/unit/resource/api1200/c7000/volume_attachment_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API1200::C7000::VolumeAttachment do
+  include_context 'shared context'
+
+  it 'inherits from OneviewSDK::API1000::C7000::VolumeAttachment' do
+    expect(described_class).to be < OneviewSDK::API1000::C7000::VolumeAttachment
+  end
+end